// src/app/_core/models/search/customer-search-vo.ts
export interface CustomerSearchVO{
    customerId:number;
    nicheIdentifier:string;
    status:number;
    categoryCustomer:string;
    companyCustomerId:number;
    beneficiariesNbr: number;
    firstName: string;
    lastName: string;
    title: string;
    crmName: string;
    companyCrmName: string;
    email: string;
    picto: string;
    libelle: string;
    univers : string;

}